added authorize for files and got rid of a few problems

+++ IssueTracker/IssueTrackerAPI/Controllers/FileController.cs
@@ -4,6 +4,8 @@ using DataAccess.Repository;
 using FluentValidation.Results;
 using IssueTracker.FileSystem;
 using IssueTrackerAPI.Utils;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Models.Request;
 using Validation;
@@ -25,8 +27,11 @@ public class FileController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public async Task<IResult> PostFile([FromForm] IFormFile formFile, [FromForm] int? issueId, [FromForm] int? commentId)
     {
+        var idclaim = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
+        if (idclaim == null) return Results.Unauthorized();
         FileRequest fileRequest = new FileRequest(formFile);
         fileRequest.IssueId = issueId;
         fileRequest.CommentId = commentId;
@@ -45,52 +50,22 @@ public class FileController : ControllerBase
             BlobName = blobFileName,
             Content = formFile.OpenReadStream(),
             SizeKb = formFile.Length / 1024,
-            Type = formFile.ContentType
+            Type = formFile.ContentType,
+            UserId = Guid.Parse(idclaim.Value)
         };
         var fileModel = new File();
         fileModel.Extension = file.Extension;
         fileModel.FileId = file.Id;
         fileModel.FileIssueId = issueId;
         fileModel.FileCommentId = commentId;
+        fileModel.FileUserId = Guid.Parse(idclaim.Value);
         await _fileRepository.AddAsync(fileModel);
         await _fileProvider.UploadAsync(file);
         return Results.Ok();
     }
 
-    [HttpGet("getFiles")]
-    public async Task<IActionResult> GetFiles(IEnumerable<FileGetRequest> fileName)
-    {
-        var fileAttachment = _mapper.Map<IEnumerable<IssueTracker.FileSystem.Models.File>>(fileName);
-        try
-        {
-            var response = await _fileProvider.GetAsync(fileAttachment);
-            if (response == null) return BadRequest("there are no files");
-            return Ok(response);
-        }
-        catch (FileSystemException ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-    [HttpGet("getone")]
-    public async Task<IActionResult> GetFile(FileGetRequest fileName)
-    {
-        var fileAttachment = _mapper.Map<IEnumerable<IssueTracker.FileSystem.Models.File>>(fileName);
-        try
-        {
-            var response = await _fileProvider.GetAsync(fileAttachment);
-
-            if (response == null) return BadRequest("Result is null!");
-            return Ok(response.FirstOrDefault());
-        }
-        catch (FileSystemException ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
     [HttpDelete]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public async Task<IResult> Delete([FromBody] FileDeleteRequest fileDelete)
     {
         var validator = new FileDeleteRequestValidation();
@@ -100,6 +75,10 @@ public class FileController : ControllerBase
             List<ValidationFailure> failures = results.Errors;
             return Results.BadRequest(failures);
         }
+        var file = await _fileRepository.GetAsync(fileDelete.FileId);
+        if (file == null) return Results.NotFound("No record to delete");
+        var idclaim = User.Claims.FirstOrDefault(x => x.Type.Equals("UserId"));
+        if (idclaim == null || file.FileUserId != Guid.Parse(idclaim.Value)) return Results.Unauthorized();
         try
         {
             await _fileRepository.DeleteAsync(fileDelete.FileId);
